Add Integration tests to check the header and lang attribute for the Ukraine in Russian IDX page
