+++ src/integration/pages/idxPage/ukraine_in_russian/amp.test.js
@@ -4,5 +4,10 @@
  */
 
 import runAmpTests from '../ampTests';
+import runHeaderTests from './header.test';
+import runFooterTests from './footer.test';
 
 describe('AMP ukrainian/ukraine_in_russian IDX page', runAmpTests);
+
+runHeaderTests();
+runFooterTests();
