+++ src/integration/pages/idxPage/ukraine_in_russian/canonical.test.js
@@ -4,5 +4,10 @@
  */
 
 import runCanonicalTests from '../canonicalTests';
+import runHeaderTests from './header.test';
+import runFooterTests from './footer.test';
 
 describe('Canonical ukrainian/ukraine_in_russian IDX page', runCanonicalTests);
+
+runHeaderTests();
+runFooterTests();
